Improve dump-eeprom as a command line utility
It can be handy to quickly see what's in the EEPROM of an EtherCAT
device, so this utility is nice.  This seeks to improve the experience
of using the utility on the command line by removing panics that are
caused by user error and some log messages that could be replaced with
pipes to other commands or aren't related to an EEPROM dump.