Don't write to 0x0980 without the user asking
Some subdevices, like my ABB FECA-01 + ACS880, will enter into a bad
state if you write to this parameter.  This removes all implicit writes
to it.

I believe that this parameter was added to Wireshark in error and
every maindevice impl has it because surely Wireshark can't be wrong.
However, it's in SOES since 2017 so there could be subdevices out there
that depend on this parameter, so it must be kept.  Let's leave it to
the user to consult their datasheet on how to use this parameter.
Especially since as far as I can tell, there are no other descriptions
of this parameter's semantics.